Attractors.jl

Attractors — Module

Attractors.jl

Attractors.jl is a Julia package for

Finding all attractors, and all types of attractors, of arbitrary dynamical systems. An extendable interface allows for new algorithms for finding attractors.

Finding their basins of attraction or the state space fractions of the basins. This includes finding exit basins (divergence to infinity).

Analyzing nonlocal stability of attractors (also called global stability or resilience).

Performing global continuation of attractors and their basins (or other measures of stability), over a parameter range. Global continuation is a new, cutting-edge type of continuation that offers several advantages over traditional local continuation (AUTO, MatCont, BifurcationKit.jl, etc.), see the comparison in our docs.

Finding the basin boundaries and edges states and analyzing their fractal properties.

Tipping points related functionality for systems with known dynamic rule.

And more!

It can be used as a standalone package, or as part of DynamicalSystems.jl.

To install it, run import Pkg; Pkg.add("Attractors").

All further information is provided in the documentation, which you can either find online or build locally by running the docs/make.jl file.

Previously, Attractors.jl was part of ChaosTools.jl

Latest news

Global continuation can now be performed across any arbitrary curve in parameter space.

See the CHANGELOG.md (at the GitHub repo) for more!

Getting started

Start by having a look at the tutorial, after which you can consult individual library functions in API. Many more examples can be found in the dedicated examples page.