Finding Attractors

Attractors.jl defines a generic interface for finding attractors of dynamical systems. One first decides the instance of GeneralizedDynamicalSystem they need. Then, an instance of AttractorMapper is created from this dynamical system. This mapper instance can be used to compute e.g., basins_of_attraction, and the output can be further analyzed to get e.g., the basin_entropy.

The example 2D basins of 4D system contains most of the functionality documented in this page.

Attractors.AttractorMapper — Type

AttractorMapper(ds::GeneralizedDynamicalSystem, args...; kwargs...) → mapper

Subtypes of AttractorMapper are structures that map initial conditions of ds to attractors. Currently available mapping methods:

AttractorsViaProximity

AttractorsViaRecurrences

AttractorsViaFeaturizing

All AttractorMapper subtypes can be used with basins_fractions or basins_of_attraction.

In addition, some mappers can be called as a function of an initial condition:

label = mapper(u0)

and this will on the fly compute and return the label of the attractor u0 converges at. The mappers that can do this are:

AttractorsViaProximity

AttractorsViaRecurrences

Proximity

Attractors.AttractorsViaProximity — Type

AttractorsViaProximity(ds::DynamicalSystem, attractors::Dict [, ε]; kwargs...)

Map initial conditions to attractors based on whether the trajectory reaches ε-distance close to any of the user-provided attractors. They have to be in a form of a dictionary mapping attractor labels to Datasets containing the attractors.

The system gets stepped, and at each step the minimum distance to all attractors is computed. If any of these distances is < ε, then the label of the nearest attractor is returned.

If an ε::Real is not provided by the user, a value is computed automatically as half of the minimum distance between all attractors. This operation can be expensive for large attractor datasets. If length(attractors) == 1, then ε becomes 1/10 of the diagonal of the box containing the attractor. If length(attractors) == 1 and the attractor is a single point, an error is thrown.

Because in this method the attractors are already known to the user, the method can also be called supervised.

Keywords

Ttr = 100: Transient time to first evolve the system for before checking for proximity.

Δt = 1: Integration step time (only valid for continuous systems).

horizon_limit = 1e3: If the maximum distance of the trajectory from any of the given attractors exceeds this limit, it is assumed that the trajectory diverged (gets labelled as -1).

mx_chk_lost = 1000: If the integrator has been stepped this many times without coming ε-near to any attractor, it is assumed that the trajectory diverged (gets labelled as -1).

diffeq = NamedTuple(): Keywords propagated to DifferentialEquations.jl (only valid for continuous systems).

Recurrences

Attractors.AttractorsViaRecurrences — Type

AttractorsViaRecurrences(ds::GeneralizedDynamicalSystem, grid::Tuple; kwargs...)

Map initial conditions to attractors by identifying attractors on the fly based on recurrences in the state space, as outlined by Datseris & Wagemakers[Datseris2022]. Works for any case encapsulated by GeneralizedDynamicalSystem.

grid is a tuple of ranges partitioning the state space so that a finite state machine can operate on top of it. For example grid = (xg, yg) where xg = yg = range(-5, 5; length = 100) for a two-dimensional system. The grid has to be the same dimensionality as the state space, use a projected_integrator if you want to search for attractors in a lower dimensional subspace.

Keyword Arguments

sparse = true: control the interval representation of the state space grid. If true, uses a sparse array, whose memory usage is in general more efficient than a regular array obtained with sparse=false. In practice, the sparse representation should always be preferred when searching for basins_fractions. Only for very low dimensional systems and for computing the full basins_of_attraction the non-sparse version should be used.

Integrator configuration

Ttr = 0: This keyword arguments allows to skip a transient before the recurrence routine begins. It is useful for some high dimensional systems to speed up the convergence to the attractor.

Δt: Approximate time step of the integrator, which is 1 for discrete time systems. For continuous systems, an automatic value is calculated using automatic_Δt_basins. For very fine grids, this can become very small, much smaller than the typical integrator internal step size in case of adaptive integrators. In that case use stop_at_Δt = true.

diffeq = NamedTuple(): Keyword arguments propagated to integrator. Only valid for continuous time systems. It is recommended to choose high accuracy solvers for this application, e.g. diffeq = (alg=Vern9(), reltol=1e-9, abstol=1e-9).

stop_at_Δt = false: control whether the integrator advances exactly Δt time points with each step or not. Should only be used if Δt is smaller than the typical integrator step (for continuous time systems). Being true slows down performance significantly but increases accuracy drastically, especially in the case of limit cycle attractors, but only if the automatic value of Δt is used.

Finite state machine configuration

mx_chk_att = 2: Μaximum checks of consecutives hits of an existing attractor cell before declaring convergence to that existing attractor.

mx_chk_hit_bas = 10: Maximum check of consecutive visits of the same basin of attraction before declaring convergence to an existing attractor.

mx_chk_fnd_att = 100: Maximum check of consecutive visits to a previously visited unlabeled cell before declaring we have found a new attractor.

mx_chk_loc_att = 100: Maximum check of consecutive visits to cells marked as a new attractor, during the attractor identification phase, before declaring we that we have identified the new attractor with sufficient cells.

store_once_per_cell = true: Control if multiple points in state space that belong to the same cell are stored or not in the attractor, after an attractor is found. If true, each visited cell will only store a point once, which is desirable for fixed points and limit cycles. If false, at least mx_chk_loc_att points are stored per attractor, leading to more densely stored attractors, which may be desirable for instance in chaotic attractors.

mx_chk_lost = 20: Maximum check of iterations outside the defined grid before we declare the orbit lost outside and hence assign it label -1.

horizon_limit = 1e6: If the norm of the integrator state reaches this limit we declare that the orbit diverged to infinity.

mx_chk_safety = Int(1e6): A safety counter that is always increasing for each initial condition. Once exceeded, the algorithm assigns -1 and throws a warning. This clause exists to stop the algorithm never haulting for innappropriate grids, where a found attractor may intersect in the same cell with a new attractor the orbit traces (which leads to infinite resetting of all counters).

Description

An initial condition given to an instance of AttractorsViaRecurrences is iterated based on the integrator corresponding to ds. A recurrence in the state space means that the trajectory has converged to an attractor. This is the basis for finding attractors.

A finite state machine (FSM) follows the trajectory in the state space, and constantly maps it to the given grid. The FSM decides when an initial condition has successfully converged into an attractor. An array, internally called "basins", stores the state of the FSM on the grid, according to the indexing system described in [Datseris2022]. As the system is integrated more and more, the information of the "basins" becomes richer and richer with more identified attractors or with grid cells that belong to basins of already found attractors. Notice that only in the special method basins_of_attraction(mapper::AttractorsViaRecurrences) the information of the attraction or exit basins is utilized. In other functions like basins_fractions only the attractor locations are utilized, as the basins themselves are not stored.

The iteration of a given initial condition continues until one of the following happens:

The trajectory hits mx_chk_fnd_att times in a row grid cells previously visited: it is considered that an attractor is found and is labelled with a new ID. Then, iteration continues a bit more until we have identified the attractor with sufficient accuracy, i.e., until mx_chk_loc_att cells with the new ID have been visited.

The trajectory hits an already identified attractor mx_chk_att consecutive times: the initial condition is numbered with the attractor's ID.

The trajectory hits a known basin mx_chk_hit_bas times in a row: the initial condition belongs to that basin and is numbered accordingly. Notice that basins are stored and used only when sparse = false.

The trajectory spends mx_chk_lost steps outside the defined grid or the norm of the integrator state becomes > than horizon_limit: the initial condition's label is set to -1.

If none of the above happens, the initial condition is labelled -1 after and mx_chk_safety integrator steps.

Missing docstring.

Missing docstring for AttractorsViaRecurrencesSparse. Check Documenter's build log for details.

Attractors.automatic_Δt_basins — Function

automatic_Δt_basins(integ, grid; N = 5000) → Δt

Calculate an optimal Δt value for basins_of_attraction. This is done by evaluating the dynamic rule f (vector field) at N randomly chosen points of the grid. The average f is then compared with the diagonal length of a grid cell and their ratio provides Δt.

Notice that Δt should not be too small which happens typically if the grid resolution is high. It is okay for basins_of_attraction if the trajectory skips a few cells. But if Δt is too small the default values for all other keywords such as mx_chk_hit_bas need to be increased drastically.

Also, Δt that is smaller than the internal step size of the integrator will cause a performance drop.

Featurizing

Attractors.AttractorsViaFeaturizing — Type

AttractorsViaFeaturizing(
    ds::DynamicalSystem, featurizer::Function,
    clusterconfig = ClusteringConfig(); kwargs...
)

Initialize a mapper that maps initial conditions to attractors using the featurizing and clustering method of [Stender2021]. See AttractorMapper for how to use the mapper.

featurizer is a function that takes as an input an integrated trajectory A::Dataset and the corresponding time vector t and returns a Vector{<:Real} of features describing the trajectory. See ClusteringConfig for configuring the clustering process.

Keyword arguments

T=100, Ttr=100, Δt=1, diffeq=NamedTuple(): Propagated to trajectory.

Description

The trajectory X of each initial condition is transformed into a vector of features. Each feature is a number useful in characterizing the attractor the initial condition ends up at, and distinguishing it from other attrators. Example features are the mean or standard deviation of one of the of the timeseries of the trajectory, the entropy of some of the dimensions, the fractal dimension of X, or anything else you may fancy. The vectors of features are then used to identify to which attractor each trajectory belongs (i.e. in which basin of attractor each initial condition is in). The method thus relies on the user having at least some basic idea about what attractors to expect in order to pick the right features, in contrast to AttractorsViaRecurrences.

Once the features are extracted, they are clustered using ClusteringConfig, so see that docstring for more details on the clustering. Each cluster is considered one attractor.

If templates are provided to ClusteringConfig, then a supervised version is used, and the functionality is similar to AttractorsViaProximity. Generally speaking, the AttractorsViaProximity is superior. However, if the dynamical system has extremely high-dimensionality, there may be reasons to use the supervised method of this featurizing algorithm instead, as it projects the trajectories into a much lower dimensional representation of features.

Attractors.ClusteringConfig — Type

ClusteringConfig(; kwargs...)

Initialize a struct that contains information used to cluster "features". These features are typically extracted from trajectories/datasets in AttractorsViaFeaturizing, or manualy by the user.

The clustering is done in the function cluster_features.

The default clustering method is an improvement over existing literature, see Description.

Keyword arguments

clust_distance_metric = Euclidean(): metric to be used in the clustering. Used in both versions.

Supervised method

templates = nothing: Enables supervised version, see below. If given (i.e. different than nothing), templates must be a Dict of cluster labels to cluster features. The labels must be of Int type, and the features are Vectors representing a cluster (which can be an attractor, for instance). The label -1 is reserved for invalid trajectories, which either diverge or whose clustering failed.

max_distance_template = Inf: Maximum allowed distance between a feature and its nearest template for it to be assigned to that template. By default, Inf guarantees that a feature is assigned to its nearest template regardless of the distance.

Unsupervised method

rescale_features = true: if true, rescale each dimension of the extracted features separately into the range [0,1]. This typically leads to more accurate clustering.

min_neighbors = 10: minimum number of neighbors (i.e. of similar features) each feature needs to have in order to be considered in a cluster (fewer than this, it is labeled as an outlier, -1).

optimal_radius_method::Union{Real, String} = "silhouettes_optim": if a real number, it is the radius used to cluster features in the unsupervised method. Otherwise, it determines the method used to automatically determine that radius. Possible values are:

"silhouettes": Performs a linear (sequential) search for the radius that maximizes a statistic of the silhouette values of clusters (typically the mean). This can be chosen with silhouette_statistic. The linear search may take some time to finish. To increase speed, the number of radii iterated through can be reduced by decreasing num_attempts_radius (see its entry below).

"silhouettes_optim": Same as "silhouettes" but performs an optimized search via Optim.jl. It's faster than "silhouettes", with typically the same accuracy (the search here is not guaranteed to always find the global maximum, though it typically gets close).

"knee": chooses the the radius according to the knee (a.k.a. elbow, highest-derivative method) and is quicker, though generally leading to much worse clustering. It requires that min_neighbors > 1.

Keywords for optimal radius estimation

num_attempts_radius = 100: number of radii that the optimal_radius_method will try out in its iterative procedure. Higher values increase the accuracy of clustering, though not necessarily much, while always reducing speed.

silhouette_statistic::Function = mean: statistic (e.g. mean or minimum) of the silhouettes that is maximized in the "optimal" clustering. The original implementation in [Stender2021] used the minimum of the silhouettes, and typically performs less accurately than the mean.

max_used_features = 0: if not 0, it should be an Int denoting the max amount of features to be used when finding the optimal radius. Useful when clustering a very large number of features (e.g., high accuracy estimation of fractions of basins of attraction).

Description

The trajectory X, which may for instance be an attractor, is transformed into a vector of features. Each feature is a number useful in characterizing the attractor, and distinguishing it from other attrators. Example features are the mean or standard deviation of one of the of the timeseries of the trajectory, the entropy of the first two dimensions, the fractal dimension of X, or anything else you may fancy. The vectors of features are then used to identify clusters of the trajectories. An application thus is to cluster attractors. Currently two methods are offered to achieve this.

Unsupervised method

The unsupervised method does not rely on templates, and instead uses the DBSCAN clustering algorithm to automatically identify clusters of similar features. To achieve this, each feature is considered a point in feature space. Each cluster then basically groups points that are closely packed together. Closely packed means that the points have at least min_neighbors inside a ball of radius optimal_radius centered on them. This method typically works well if the radius is chosen well, which is not necessarily an easy task. Currently, three methods are implemented to automatically estimate an "optimal" radius.

Estimating the optimal radius

The default method is the silhouettes method, which includes keywords silhouette and silhouette_optim. Both of them search for the radius that optimizes the clustering, meaning the one that maximizes a statistic silhouette_statistic (e.g. mean value) of a quantifier for the quality of each cluster. This quantifier is the silhouette value of each identified cluster. A silhouette value measures how similar a point is to the cluster it currently belongs to, compared to the other clusters, and ranges from -1 (worst matching) to +1 (ideal matching). If only one cluster is found, the assigned silhouette is

So for each attempted radius in the search the clusters are computed, their silhouettes

calculated, and the statistic of these silhouettes computed. The algorithm then finds the radius that leads to the maximum such statistic. For optimal_radius_method = "silhouettes", the search is done linearly, from a minimum to a maximum candidate radius for optimal_radius_method = "silhouettes"; optimal_radius_method = silhouettes_optim, it is done via an optimized search performed by Optim.jl which is typically faster and with similar accuracy. A third alternative is the"elbow" method, which works by calculating the distance of each point to its k-nearest-neighbors (with k=min_neighbors) and finding the distance corresponding to the highest derivative in the curve of the distances, sorted in ascending order. This distance is chosen as the optimal radius. It is described in [Kriegel1996] and [Schubert2017]. It typically performs considerably worse than the "silhouette" methods.

Supervised method

In the alternative, supervised version, the user provides features to be used as templates guiding the clustering via the templates keyword. Features are clustered to their nearest template if the distance to that template is below or equal to the threshold max_distance_template. The distances are computed via the metric defined by clust_distance_metric. So the algorithm first finds the template nearest to each feature (in that metric) via the k-nearest-neighbors algorithm for k=1 and then compares their distance to the threshold. If the distance is below or equal to the threshold, the feature is assigned the template's label, which is given in templates. If the distance is larger, the feature is assigned label -1.

Attractors.cluster_features — Function

cluster_features(features, cc::ClusteringConfig)

Cluster the given features::Vector{<:AbstractVector}, according to given ClusteringConfig. Return cluster_labels, which contains, for each feature, the labels (indices) of the corresponding cluster.

Datseris2022G. Datseris and A. Wagemakers, Effortless estimation of basins of attraction, Chaos 32, 023104 (2022)

Stender2021Stender & Hoffmann, bSTAB: an open-source software for computing the basin stability of multi-stable dynamical systems

Stender2021Stender & Hoffmann, bSTAB: an open-source software for computing the basin stability of multi-stable dynamical systems

Kriegel1996Ester, Kriegel, Sander and Xu: A Density-Based Algorithm for Discovering Clusters in Large Spatial Databases with Noise

Schubert2017Schubert, Sander, Ester, Kriegel and Xu: DBSCAN Revisited, Revisited: Why and How You Should (Still) Use DBSCAN