Finding Attractors

Attractors.jl defines a generic interface for finding attractors of dynamical systems. One first decides the instance of GeneralizedDynamicalSystem they need. Then, an instance of AttractorMapper is created from this dynamical system. This mapper instance can be used to compute e.g., basins_of_attraction, and the output can be further analyzed to get e.g., the basin_entropy.

The example 2D basins of 4D system contains most of the functionality documented in this page.

Attractors.AttractorMapper — Type

AttractorMapper(ds::GeneralizedDynamicalSystem, args...; kwargs...) → mapper

Subtypes of AttractorMapper are structures that map initial conditions of ds to attractors. Currently available mapping methods:

AttractorsViaProximity

AttractorsViaRecurrences

AttractorsViaFeaturizing

All AttractorMapper subtypes can be used with basins_fractions or basins_of_attraction.

In addition, some mappers can be called as a function of an initial condition:

label = mapper(u0)

and this will on the fly compute and return the label of the attractor u0 converges at. The mappers that can do this are:

AttractorsViaProximity

AttractorsViaRecurrences

Proximity

Attractors.AttractorsViaProximity — Type

AttractorsViaProximity(ds::DynamicalSystem, attractors::Dict [, ε]; kwargs...)

Map initial conditions to attractors based on whether the trajectory reaches ε-distance close to any of the user-provided attractors. They have to be in a form of a dictionary mapping attractor labels to Datasets containing the attractors.

The system gets stepped, and at each step the minimum distance to all attractors is computed. If any of these distances is < ε, then the label of the nearest attractor is returned.

If an ε::Real is not provided by the user, a value is computed automatically as half of the minimum distance between all attractors. This operation can be expensive for large attractor datasets. If length(attractors) == 1, then ε becomes 1/10 of the diagonal of the box containing the attractor. If length(attractors) == 1 and the attractor is a single point, an error is thrown.

Because in this method the attractors are already known to the user, the method can also be called supervised.

Keywords

Ttr = 100: Transient time to first evolve the system for before checking for proximity.

Δt = 1: Integration step time (only valid for continuous systems).

horizon_limit = 1e3: If the maximum distance of the trajectory from any of the given attractors exceeds this limit, it is assumed that the trajectory diverged (gets labelled as -1).

mx_chk_lost = 1000: If the integrator has been stepped this many times without coming ε-near to any attractor, it is assumed that the trajectory diverged (gets labelled as -1).

diffeq = NamedTuple(): Keywords propagated to DifferentialEquations.jl (only valid for continuous systems).

Recurrences

Attractors.AttractorsViaRecurrences — Type

AttractorsViaRecurrences(ds::GeneralizedDynamicalSystem, grid::Tuple; kwargs...)

Map initial conditions to attractors by identifying attractors on the fly based on recurrences in the state space, as outlined by Datseris & Wagemakers[Datseris2022]. Works for any case encapsulated by GeneralizedDynamicalSystem. The version AttractorsViaRecurrencesSparse should practically always be preferred over this one.

grid is a tuple of ranges partitioning the state space so that a finite state machine can operate on top of it. For example grid = (xg, yg) where xg = yg = range(-5, 5; length = 100) for a two-dimensional system. The grid has to be the same dimensionality as the state space, use a projected_integrator if you want to search for attractors in a lower dimensional subspace.

Keyword Arguments

Δt: Approximate time step of the integrator, which is 1 for discrete systems. For continuous systems, an automatic value is calculated using automatic_Δt_basins.

Ttr = 0: This keyword arguments allows to skip a transient before the recurrence routine begins. It is useful for some high dimensional systems to speed up the convergence to the attractor.

diffeq = NamedTuple(): Keyword arguments propagated to integrator. Only valid for ContinuousDynamicalSystem. It is recommended to choose high accuracy solvers for this application, e.g. diffeq = (alg=Vern9(), reltol=1e-9, abstol=1e-9).

mx_chk_att = 2: A parameter that sets the maximum checks of consecutives hits of an attractor before deciding the basin of the initial condition.

mx_chk_hit_bas = 10: Maximum check of consecutive visits of the same basin of attraction. This number can be increased for higher accuracy.

mx_chk_fnd_att = 100: Maximum check of unnumbered cell before considering we have an attractor. This number can be increased for higher accuracy.

mx_chk_loc_att = 100: Maximum check of consecutive cells marked as an attractor before considering that we have all the available pieces of the attractor.

mx_chk_lost = 20: Maximum check of iterations outside the defined grid before we consider the orbit lost outside. This number can be increased for higher accuracy.

horizon_limit = 1e6: If the norm of the integrator state reaches this limit we consider that the orbit diverges.

safety_counter_max = Int(1e6): A safety counter that is always increasing for each initial condition. Once exceeded, the algorithm assigns -1 and throws a warning. This clause exists to stop the algorithm never haulting for innappropriately defined grids, where a found attractor may intersect in the same cell with a new attractor the orbit traces (which leads to infinite resetting of all counters). As this check comes with a performance deficit, the keyword unsafe=true can be set to disable it in case the user is confident the algorithm will hault.

Description

An initial condition given to an instance of AttractorsViaRecurrences is iterated based on the integrator corresponding to ds. A recurrence in the state space means that the trajectory has converged to an attractor. This is the basis for finding attractors.

A finite state machine (FSM) follows the trajectory in the state space, and constantly maps it to the given grid. The FSM decides when an initial condition has successfully converged into an attractor. An array, internally called "basins", stores the state of the FSM on the grid, according to the indexing system described in [Datseris2022]. As the system is integrated more and more, the information of the "basins" becomes richer and richer with more identified attractors or with grid cells that belong to basins of already found attractors. Notice that only in the special method basins_of_attraction(mapper::AttractorsViaRecurrences) the information of the attraction or exit basins is utilized. In other functions like basins_fractions only the attractor locations are utilized.

The iteration of a given initial condition continues until one of the following happens:

The trajectory hits mx_chk_fnd_att times in a row grid cells previously visited: it is considered that an attractor is found and is labelled with a new number.

The trajectory hits an already identified attractor mx_chk_att consecutive times: the initial condition is numbered with the attractor's number.

The trajectory hits a known basin mx_chk_hit_bas times in a row: the initial condition belongs to that basin and is numbered accordingly.

The trajectory spends mx_chk_lost steps outside the defined grid or the norm of the integrator state becomes > than horizon_limit: the initial condition is set to -1.

If none of the above happens and unsafe=true, the algorithm will error.

Attractors.AttractorsViaRecurrencesSparse — Function

AttractorsViaRecurrencesSparse(ds::GeneralizedDynamicalSystem, grid::Tuple; kwargs...)

This version is practically identical to AttractorsViaRecurrences, with the difference that the internal representation of the grid uses a sparse array. In practice, it should always be preferred when searching for basins_fractions. Only for very low dimensional systems and for computing the full basins_of_attraction the non-sparse version should be used.

See the docstring of AttractorsViaRecurrences for possible keywords and details on the algorithm.

Attractors.automatic_Δt_basins — Function

automatic_Δt_basins(integ, grid; N = 5000) → Δt

Calculate an optimal Δt value for basins_of_attraction. This is done by evaluating the dynamic rule f (vector field) at N randomly chosen points of the grid. The average f is then compared with the diagonal length of a grid cell and their ratio provides Δt.

Notice that Δt should not be too small which happens typically if the grid resolution is high. It is okay for basins_of_attraction if the trajectory skips a few cells. But if Δt is too small the default values for all other keywords such as mx_chk_hit_bas need to be increased drastically.

Also, Δt that is smaller than the internal step size of the integrator will cause a performance drop.

Featurizing

Attractors.AttractorsViaFeaturizing — Type

AttractorsViaFeaturizing(
    ds::DynamicalSystem, featurizer::Function,
    clusterconfig = ClusteringConfig(); kwargs...
)

Initialize a mapper that maps initial conditions to attractors using the featurizing and clustering method of [Stender2021]. See AttractorMapper for how to use the mapper.

featurizer is a function that takes as an input an integrated trajectory A::Dataset and the corresponding time vector t and returns a Vector{<:Real} of features describing the trajectory. See ClusteringConfig for configuring the clustering process.

Keyword arguments

T=100, Ttr=100, Δt=1, diffeq=NamedTuple(): Propagated to trajectory.

Description

The trajectory X of each initial condition is transformed into a vector of features. Each feature is a number useful in characterizing the attractor the initial condition ends up at, and distinguishing it from other attrators. Example features are the mean or standard deviation of one of the of the timeseries of the trajectory, the entropy of some of the dimensions, the fractal dimension of X, or anything else you may fancy. The vectors of features are then used to identify to which attractor each trajectory belongs (i.e. in which basin of attractor each initial condition is in). The method thus relies on the user having at least some basic idea about what attractors to expect in order to pick the right features, in contrast to AttractorsViaRecurrences.

Once the features are extracted, they are clustered using ClusteringConfig, so see that docstring for more details on the clustering. Each cluster is considered one attractor.

If templates are provided to ClusteringConfig, then a supervised version is used, and the functionality is similar to AttractorsViaProximity. Generally speaking, the AttractorsViaProximity is superior. However, if the dynamical system has extremely high-dimensionality, there may be reasons to use the supervised method of this featurizing algorithm instead, as it projects the trajectories into a much lower dimensional representation of features.

Attractors.ClusteringConfig — Type

ClusteringConfig(; kwargs...)

Initialize a struct that contains information used to cluster "features". These features are typically extracted from trajectories/datasets in AttractorsViaFeaturizing, or manualy by the user.

The clustering is done in the function cluster_features.

The default clustering method is an improvement over existing literature, see Description.

Keyword arguments

templates = nothing: Enables supervised version, see below. If given, must be a Dict of cluster labels to cluster features. The labels must be of Int type, and the features are Vectors representing a cluster (which can be an attractor, for instance). The label -1 is reserved for invalid trajectories, which either diverge or whose clustering failed.

Supervised method

clustering_threshold = 0.0: Maximum allowed distance between a feature and the cluster center for it to be considered inside the cluster. Only used when clust_method = "kNN_thresholded".

clust_method = clustering_threshold > 0 ? "kNN_thresholded" : "kNN": (supervised method only) which clusterization method to apply. If "kNN", the first-neighbor clustering is used. If "kNN_thresholded", a subsequent step is taken, which considers as unclassified (label -1) the features whose distance to the nearest template is above the clustering_threshold.

Unsupervised method

rescale_features = true: if true, rescale each dimension of the extracted features separately into the range [0,1]. This typically leads to more accurate clustering.

min_neighbors = 10: (unsupervised method only) minimum number of neighbors (i.e. of similar features) each feature needs to have in order to be considered in a cluster (fewer than this, it is labeled as an outlier, -1).

clust_method_norm = Euclidean(): metric to be used in the clustering.

optimal_radius_method::String = "silhouettes_optim": the method used to determine the optimal radius for clustering features in the unsupervised method. Possible values are:

"silhouettes": Performs a linear (sequential) search for the radius that maximizes a statistic of the silhouette values of clusters (typically the mean). This can be chosen with silhouette_statistic. The linear search may take some time to finish. To increase speed, the number of radii iterated through can be reduced by decreasing num_attempts_radius (see its entry below).

"silhouettes_optim": Same as "silhouettes" but performs an optimized search via Optim.jl. It's faster than "silhouettes", with typically the same accuracy (the search here is not guaranteed to always find the global maximum, though it typically gets close).

"knee": chooses the the radius according to the knee (a.k.a. elbow, highest-derivative method) and is quicker, though generally leading to much worse clustering. It requires that min_neighbors > 1.

Keywords for optimal radius estimation

num_attempts_radius = 100: number of radii that the optimal_radius_method will try out in its iterative procedure. Higher values increase the accuracy of clustering, though not necessarily much, while always reducing speed.

silhouette_statistic::Function = mean: statistic (e.g. mean or minimum) of the silhouettes that is maximized in the "optimal" clustering. The original implementation in [Stender2021] used the minimum of the silhouettes, and typically performs less accurately than the mean.

max_used_features = 0: if not 0, it should be an Int denoting the max amount of features to be used when finding the optimal radius. Useful when clustering a very large number of features (e.g., high accuracy estimation of fractions of basins of attraction).

Description

The trajectory X, which may for instance be an attractor, is transformed into a vector of features. Each feature is a number useful in characterizing the attractor, and distinguishing it from other attrators. Example features are the mean or standard deviation of one of the of the timeseries of the trajectory, the entropy of the first two dimensions, the fractal dimension of X, or anything else you may fancy. The vectors of features are then used to identify clusters of attractors.

There are two versions to do this. The unsupervised version does not rely on templates, and instead uses the DBSCAN clustering algorithm to identify clusters of similar features. To achieve this, each feature is considered a point in feature space. In this space, the algorithm basically groups points that are closely packed. To achieve this, a crucial parameter is a radius for distance ϵ that determines the "closeness" of points in clusters.

Two methods are currently implemented to determine an optimal_radius, as described in optimal_radius_method above. The default method is based on the silhouette values of clusters. A silhouette value measures how similar a point is to the cluster it currently belongs, compared to the other clusters, and ranges from -1 (worst matching) to +1 (ideal matching). If only one cluster is found, the assigned silhouette is 0. The default method ,silhouettes, chooses the radius ε that maximizes the average silhouette across all clusters. The alternative elbow method works by calculating the distance of each point to its k-nearest-neighbors (with k=min_neighbors) and finding the distance corresponding to the highest derivative in the curve of the distances, sorted in ascending order. This distance is chosen as the optimal radius. It is described in [Kriegel1996] and [Schubert2017].

In the supervised version, the user provides features to be used as templates guiding the clustering via the templates keyword. Each feature is considered to belong to the "cluster" of the nearest template (based on the distance in feature space), and is labelled following the template's label, given in templates.

In both versions, features whose cluster is not identified are labeled as -1.

Attractors.cluster_features — Function

cluster_features(features, cc::ClusteringConfig)

Cluster the given features::Vector{<:AbstractVector}, according to given ClusteringConfig. Return cluster_labels, cluster_errors, which respectively contain, for each feature, the labels (indices) of the corresponding cluster and the error associated with that clustering. The error is the distance from the feature to (i) the cluster, in the supervised method or (ii) to the center of the cluster, in the unsupervised method.

Datseris2022G. Datseris and A. Wagemakers, Effortless estimation of basins of attraction, Chaos 32, 023104 (2022)

Stender2021Stender & Hoffmann, bSTAB: an open-source software for computing the basin stability of multi-stable dynamical systems

Stender2021Stender & Hoffmann, bSTAB: an open-source software for computing the basin stability of multi-stable dynamical systems

Kriegel1996Ester, Kriegel, Sander and Xu: A Density-Based Algorithm for Discovering Clusters in Large Spatial Databases with Noise

Schubert2017Schubert, Sander, Ester, Kriegel and Xu: DBSCAN Revisited, Revisited: Why and How You Should (Still) Use DBSCAN