Attractor & Basins Continuation

Basins continuation API

Attractors.basins_fractions_continuation — Function

basins_fractions_continuation(continuation, prange, pidx, ics; kwargs...)

Find and continue attractors and the fractions of their basins of attraction across a parameter range. The given continuation contains a reference to a dynamical system, as well as how to find its attractors. I.e., it contains an AttractorMapper. Given this continuation, the basin fractions (and the attractors for the RecurrencesSeedingContinuation method) are continued across the parameter range prange, for the parameter of the system with index pidx. ics is as in basins_fractions, i.e., it is either a function generating initial conditions or a dataset containing them.

Return

fractions_curves <: Vector{Dict{Int, Float64}}. The fractions of basins of attraction. fractions_curves[i] is a dictionary mapping attractor IDs to their basin fraction at the i-th parameter.

attractors_info <: Vector{Dict{Int, <:Any}}. Information about the attractors. attractors_info[i] is a dictionary mapping attractor ID to information about the attractor at the i-th parameter. The type of information stored depends on the chosen continuation method.

Keyword arguments

samples_per_parameter = 100: Amount of initial conditions sampled at each parameter.

show_progress = true: Whether to show a progress bar for the process.

Continuation methods

RecurrencesSeedingContinuation. For this sampler, ics is optional. If not given, one is created using the grid of AttractorsViaRecurrences:
sampler, = statespace_sampler(min_bounds = minimum.(grid), max_bounds = maximum.(grid))

GroupAcrossParameterContinuation.

Recurrences continuation (best)

Attractors.RecurrencesSeedingContinuation — Type

RecurrencesSeedingContinuation(mapper::AttractorsViaRecurrences; kwargs...)

A method for basins_fractions_continuation. It uses seeding of previous attractors to find new ones, which is the main performance bottleneck. The method uses match_attractor_ids! to match attractors as the system parameter is increased.

Will write more once we have the paper going.

Keyword Arguments

method, threshold: Given to match_attractor_ids! which is the function used to match attractors between each parameter slice.

info_extraction = identity: A function that takes as an input an attractor (Dataset) and outputs whatever information should be stored. It is used to return the attractors_info in basins_fractions_continuation.

seeds_from_attractor: A function that takes as an input an attractor and returns an iterator of initial conditions to be seeded from the attractor for the next parameter slice. By default, we sample some points from existing attractors according to how many points the attractors themselves contain. A maximum of 10 seeds is done per attractor.

Matching attractors

Attractors.match_attractor_ids! — Function

match_attractor_ids!(a₊::AbstractDict, a₋; method = Centroid(), threshold = Inf)

Given dictionaries a₊, a₋ mapping IDs to attractors (Dataset instances), match attractor IDs in dictionary a₊ so that its attractors that are the closest to those in dictionary a₋ get assigned the same key as in a₋. Typically the +,- mean after and before some change of parameter of a system.

Return the replacement map, a dictionary mapping old keys of a₊ to the new ones that they were mapped to. You can obtain this map, without modifying the dictionaries, by directly calling the replacement_map function directly.

Description

When finding attractors and their fractions in DynamicalSystems.jl, different attractors get assigned different IDs. However which attractor gets which ID is somewhat arbitrary. Finding the attractors of the same system for slightly different parameters could label "similar" attractors (at the different parameters) with different IDs. match_attractors_ids! tries to "match" them by modifying the attractor IDs, i.e., the keys of the given dictionaries.

The matching happens according to the output of the setsofsets_distance function with the keyword method. method` can be whatever that function accepts. Attractors are then match according to distance, with unique mapping. The closest attractors (before and after) are mapped to each other, and are removed from the matching pool, and then the next pair with least remaining distance is matched, and so on.

Additionally, you can provide a threshold value. If the distance between two attractors is larger than this threshold, then it is guaranteed that the attractors will get assigned different key in the dictionary a₊.

Attractors.replacement_map — Function

replacement_map(a₊, a₋; method = Centroid(), threshold = Inf) → rmap

Return a dictionary mapping keys in a₊ to new keys in a₋, as explained in match_attractor_ids!.

Missing docstring.

Missing docstring for dataset_distance. Check Documenter's build log for details.

StateSpaceSets.setsofsets_distance — Function

setsofsets_distance(a₊, a₋ [, method]) → distances

Calculate distances between sets of StateSpaceSets. Here a₊, a₋ are containers of StateSpaceSets, and the returned distances are dictionaries of of distances. Specifically, distances[i][j] is the distance of the set in the i key of a₊ to the j key of a₋. Notice that distances from a₋ to a₊ are not computed at all (assumming symmetry in the distance function).

The method can be as in set_distance. However, method can also be any arbitrary user function that takes as input two datasets and returns any positive-definite number as their "distance".

Attractors.match_basins_ids! — Function

match_basins_ids!(b₊::AbstractArray, b₋; threshold = Inf)

Similar to match_attractor_ids! but operate on basin arrays instead (the arrays typically returned by basins_of_attraction).

This method matches IDs of attractors whose basins of attraction before and after b₋,b₊ have the most overlap (in pixels). This overlap is normalized in 0-1 (with 1 meaning 100% overlap of pixels). The threshold in this case is compared to the inverse of the overlap (so, for threshold = 2 attractors that have less than 50% overlap get different IDs guaranteed).

Aggregating attractors and fractions

Attractors.aggregate_attractor_fractions — Function

aggregate_attractor_fractions(
    fractions_curves, attractors_info, featurizer, group_config [, info_extraction]
)

Aggregate the already-estimated curves of fractions of basins of attraction of similar attractors using the same pipeline used by GroupingConfig. The most typical application of this function is to transform the output of RecurrencesSeedingContinuation so that similar attractors, even across parameter space, are grouped into one "attractor". Thus, the fractions of their basins are aggregated.

You could also use this function to aggregate attractors and their fractions even in a single parameter configuration, i.e., using the output of basins_fractions.

This function is useful in cases where you want the accuracy and performance of AttractorsViaRecurrences, but you also want the convenience of "grouping" similar attractrors like in AttractorsViaFeaturizing for presentation or analysis purposes. For example, a high dimesional model of competition dynamics across multispecies may have extreme multistability. After finding this multistability However, one may care about aggregating all attractors into two groups: where a given species is extinct or not. This is the example highlighted in our documentation, in Extinction of a species in a multistable competition model.

Input

fractions_curves: a vector of dictionaries mapping labels to basin fractions.

attractors_info: a vector of dictionaries mapping labels to attractors. 1st and 2nd argument are exactly like the return values of basins_fractions_continuation with RecurrencesSeedingContinuation (or, they can be the return of basins_fractions).

featurizer: a 1-argument function to map an attractor into a feature SVector. Notice that you can use identity if the input "attractors" aren't actually attractors but already features or something else that can be grouped directly.

group_config: a subtype of GroupingConfig.

info_extraction: a function accepting a vector of features and returning a description of the features. I.e., exactly as in GroupAcrossParameterContinuation. The 5th argument is optional and defaults to the centroid of the features.

Return

aggregated_fractions: same as fractions_curves but now contains the fractions of the aggregated attractors.

aggregated_info: dictionary mapping the new labels of aggregated_fractions to the extracted information using info_extraction.

Grouping continuation

Attractors.GroupAcrossParameterContinuation — Type

GroupAcrossParameterContinuation(mapper::AttractorsViaFeaturizing; kwargs...)

A method for basins_fractions_continuation. It uses the featurizing approach discussed in AttractorsViaFeaturizing and hence requires an instance of that mapper as an input. When used in basins_fractions_continuation, features are extracted and then grouped across a parameter range. Said differently, all features of all initial conditions across all parameter values are put into the same "pool" and then grouped as dictated by the group_config of the mapper. After the grouping is finished the feature label fractions are distributed to each parameter value they came from.

Keyword arguments

info_extraction::Function a function that takes as an input a vector of features (corresponding to a cluster) and returns a description of the cluster. By default, the centroid of the cluster is used.

par_weight = 0: See below the section on MCBB.

MCBB special version

If the chosen grouping method is GroupViaClustering, the additional keyword par_weight::Real can be used. If it is ≠ 0, the distance matrix between features obtains an extra weight that is proportional to the distance par_weight*|p[i] - p[j]| between the parameters used when extracting features. The range of parameters is normalized to 0-1 such that the largest distance in the parameter space is 1. The normalization is done because the feature space is also (by default) normalized to 0-1.

This version of the algorithm is the original "MCBB" continuation method described in [Gelbrecht2020], besides the improvements of clustering accuracy and performance done by the developer team of Attractors.jl.

Gelbrecht2021Maximilian Gelbrecht et al 2021, Monte Carlo basin bifurcation analysis, New J. Phys.22 03303