Attractor & Basins Continuation

Matching

Attractors.match_attractor_ids! — Function

match_attractor_ids!(a₊::AbstractDict, a₋; metric = Euclidean(), threshold = Inf)

Given dictionaries a₊, a₋ mapping IDs to attractors (Dataset instances), match attractor IDs in dictionary a₊ so that its attractors that are the closest to those in dictionary a₋ get assigned the same key as in a₋. Typically the +,- mean after and before some change of parameter of a system.

Return the replacement map, a dictionary mapping old keys of a₊ to the new ones that they were mapped to. You can obtain this map, without modifying the dictionaries, by directly calling the replacement_map function directly.

Description

When finding attractors and their fractions in DynamicalSystems.jl, different attractors get assigned different IDs. However which attractor gets which ID is somewhat arbitrary. Finding the attractors of the same system for slightly different parameters could label "similar" attractors (at the different parameters) with different IDs. match_attractors_ids! tries to "match" them by modifying the attractor IDs, i.e., the keys of the given dictionaries.

The matching happens according to the output of the datasets_sets_distances function with the keyword metric. metric can be whatever that function accepts, such as an actual Metric instance, or an arbitrary user-defined function that computes an arbitrary "distance" between two datasets. Attractors are then match according to distance, with unique mapping. The closest attractors (before and after) are mapped to each other, and are removed from the matching pool, and then the next pair with least remaining distance is matched, and so on.

Additionally, you can provide a threshold value. If the distance between two attractors is larger than this threshold, then it is guaranteed that the attractors will get assigned different key in the dictionary a₊.

Attractors.replacement_map — Function

replacement_map(a₊, a₋; metric = Euclidean(), threshold = Inf) → rmap

Return a dictionary mapping keys in a₊ to new keys in a₋, as explained in match_attractor_ids!.

DelayEmbeddings.dataset_distance — Function

dataset_distance(dataset1, dataset2 [, method])

Calculate a distance between two AbstractDatasets, i.e., a distance defined between sets of points, as dictated by method. method defaults to Euclidean().

Description

If method isa Metric from Distances.jl, then the distance is the minimum distance of all the distances from a point in one set to the closest point in the other set. The distance is calculated with the given metric. In this case there is an internal heuristic: if length(dataset1)*length(dataset2) ≥ 1000 the algorithm switches to a KDTree-based version, otherwise it uses brute force. You can overwrite this by explicitly setting the brute boolean keyword or calling dataset_distance_brute directly.

method can also be Hausdorff (a name provided by this module).

DelayEmbeddings.datasets_sets_distances — Function

datasets_sets_distances(a₊, a₋ [, metric/method]) → distances

Calculate distances between sets of Datasets. Here a₊, a₋ are containers of Datasets, and the returned distances are dictionaries of of distances. Specifically, distances[i][j] is the distance of the dataset in the i key of a₊ to the j key of a₋. Notice that distances from a₋ to a₊ are not computed at all (assumming symmetry in the distance function).

The metric/method can be as in dataset_distance, in which case both sets must have equal-dimension datasets. However, method can also be any arbitrary user function that takes as input two datasets and returns any positive-definite number as their "distance".

Attractors.match_basins_ids! — Function

match_basins_ids!(b₊::AbstractArray, b₋; threshold = Inf)

Similar to match_attractor_ids! but operate on basin arrays instead (the arrays typically returned by basins_of_attraction).

This method matches IDs of attractors whose basins of attraction before and after b₋,b₊ have the most overlap (in pixels). This overlap is normalized in 0-1 (with 1 meaning 100% overlap of pixels). The threshold in this case is compared to the inverse of the overlap (so, for threshold = 2 attractors that have less than 50% overlap get different IDs guaranteed).

Continuation

Attractors.basins_fractions_continuation — Function

basins_fractions_continuation(continuation, prange, pidx, ics; kwargs...)

The given continuation contains a reference to a dynamical system, as well as how to find its attractors. I.e., it contains an AttractorMapper. Given this continuation, the basin fractions of the dynamical system, and continuate them across the parameter range prange, for the parameter of the system with index pidx. ics is as in basins_fractions, i.e., it must be a function generating random initial conditions.

Current continuation methods are:

RecurrencesSeedingContinuation.

Return:

fractions_curves <: Vector{Dict{Int, Float64}}. The fractions of basins of attraction. fractions_curves[i] is a dictionary mapping attractor IDs to their basin fraction at the i-th parameter.

attractors_info <: Vector{Dict{Int, <:Any}}. Information about the attractors. attractors_info[i] is a dictionary mapping attractor ID to information about the attractor at the i-th parameter. The type of information stored depends on the chosen continuation method.

The keyword arguments accepted are:

samples_per_parameter = 100: Amount of random samples drawn for each parameter value.

show_progress = true: Whether to show a progress bar for the process.

Attractors.RecurrencesSeedingContinuation — Type

RecurrencesSeedingContinuation(mapper::AttractorsViaRecurrences; kwargs...)

A method for basins_fractions_continuation. It uses seeding of previous attractors to find new ones, which is the main performance bottleneck. Will write more once we have the paper going.

Keyword Arguments

metric, threshold: Given to match_attractor_ids! which is the function used to match attractors between each parameter slice.

info_extraction = identity: A function that takes as an input an attractor (Dataset) and outputs whatever information should be stored. It is used to return the attractors_info in basins_fractions_continuation.

seeds_from_attractor: A function that takes as an input an attractor and returns an iterator of initial conditions to be seeded from the attractor for the next parameter slice. By default, we sample some points from existing attractors according to how many points the attractors themselves contain. A maximum of 10 seeds is done per attractor.